revise contract translator and fix typo
